+++ Tests/VCoreTests/Tests/Extensions/UIKit/UIImageCompressedTests.swift
@@ -12,9 +12,9 @@ import XCTest
 
 // MARK: - Tests
 final class UIImageCompressedTests: XCTestCase {
-    func test() {
+    func testJPEG() {
         let image: UIImage = .init(size: .init(dimension: 100), color: .red)! // Force-unwrap
-        let compressedImage: UIImage = image.compressed(quality: 0.75)! // Force-unwrap
+        let compressedImage: UIImage = image.jpegCompressed(quality: 0.75)! // Force-unwrap
         
         let imageData: Data = image.jpegData(compressionQuality: 1)! // Force-unwrap
         let compressedImageData: Data = compressedImage.jpegData(compressionQuality: 1)! // Force-unwrap